Prototype hack: Suffix Sort, for extending pathkeys
If you have an index on a, this patch will allow you to use it to implement
ORDER BY a, b.  That avoids a big sort up front.  Costing changes have not
been done yet so you may need to disable seqscan to reach the index scan
plan...

This is quick and dirty experiment grade code.