Remove UndoDiscard and move its state into UndoLogControl.
Previously, a per-undo log object "UndoDiscard" was used to track the progress
of the undo worker machinery.  There were some implementation problems with
it, including the use of a global array of these objects whose fixed size was
not going to work.  On closer inspection the separate UndoDiscard state seemed
to be unnecessary.  This commit switches to using the discard pointer
in the existing per-undo log UndoLogControl object, and moves the discard lock
and the oldest_xid in there too.

Since this change requires undodiscard.c to have direct access to
UndoLogControl objects and to iterate over them, create new functions
UndoLogNext(), UndoLogGet() with extern linkage to do that.  A better
interface is probably needed -- to review later when we work out the type of
access that multi-process undo worker infrastructure will need.

Thomas Munro, reviewed by Dilip Kumar and Amit Kapila