Fix uur_next chain by moving global variables into UndoLogControl.
Previously, global variables tracked information about the most recently
inserted transaction.  That wasn't appropriate because you could switch to
a different undo log at any time and (in future commits) you could be attached
to more than one undo log at a time.  The resulting corruption would cause
the undo worker to crash.

Since this is really per-undo log, let's move prev_xid into UndoLogControl
where it can be accessed without locking by the backend that is attached to
the undo log, and used the pre-existing last_xact_start and prevlen members
of UndoLogMeta which can be read (but not written) by the backend that is
attached.  There was no benefit to maintaining a backend-local cache of
that information, and we get to delete some global variables.

Thomas Munro, reviewed by XXX