Add a smgrsync() implementation for undofile.c.
Teach the checkpointer to go through an sgmr API call to mark a (RelFileNode,
forknum, segment) as in need of flushing to disk at the next call to
smgrsync().  Previously it called RememberFsyncRequest() in md.c directly,
but undofile.c needs to be able to participate in this scheme too.  So, add a
new function smgrrequestsync(), and have it forward to mdrequestsync() or
undofile_requestsync() as appropriate.

For now, there is a LOG message when undo segment files are fsync'd, like the
existing create/recycle/unlink messages.  These will be removed in future.

WIP:  undofile_sync() and register_dirty_segment copy a few too many lines
from mdsync().  Also there is a duplicated translation-file private
pendingOpsTable.  Contemplating ways to fix these problems.

Thomas Munro, RM43460