Exit by default if postmaster dies.
Users of the WaitEventSet and WaitLatch() APIs can now choose between asking
for W_POSTMASTER_DEATH explicitly and then handling it, or not asking for it
and getting the new default behavior.

Update call sites to use the new simplified interface.  In all cases where
the behavior was to exit(1) or proc_exit(1), remove that code and rely on the
default behavior.  In some cases where WL_POSTMASTER_DEATH was requested but
apparently ignored, also use the new default behavior because that was
probably a bug.  The remaining cases of explicit WL_POSTMASTER_DEATH events
are retained because they do something special when the postmaster dies.

Author: Thomas Munro
Discussion: https://postgr.es/m/CAEepm%3D1TCviRykkUb69ppWLr_V697rzd1j3eZsRMmbXvETfqbQ%40mail.gmail.com,
            https://postgr.es/m/CAEepm=2LqHzizbe7muD7-2yHUbTOoF7Q+qkSD5Q41kuhttRTwA@mail.gmail.com