Support different tuple locking modes
The basic idea is that we maintain each lockers information in undo
and the strongest lockers information on the tuple. If there is more
than one locker, then we set multi_locker bit on tuple. Now, if the
multi_locker bit is set and the new locker conflicts with the
strongest locker, then we traverse all the undo chains in the page and
wait for all the conflicting lockers to finish. As we have to wait for
all the lockers by releasing the lock on the buffer and then reacquire
the buffer lock after waiting for all the transactions is finished, in
the meantime, a new locker (say key share) can take a lock on the
tuple and we won't be able to detect it unless we do something
special.  Now, one might think that as before waiting for multiple
lockers we have acquired a heavyweight lock on tuple by using
heap_acquire_tuplock, no other transaction can acquire xid-based lock
(something like key share), but that is not true, as that is allowed for
both heap as well as for zheap (till now).

Now, the heap can detect such a case because it always creates a new
multixact whenever a new locker is added to the existing set of
lockers and it puts the newly create multixact id in xmax of tuple, so
in above case after reacquiring the buffer lock we can just check if
the xmax has changed and if so, then we redo the *TupleSatisfies check
and again wait for new lockers.

For zheap, after reacquiring the buffer lock, check again if there is
any new locker on the tuple and to find this we need to again traverse
the undo chains. Although this doesn't sound the best design, it is
not clear whether it can really create the problem. If we want we can
optimize by checking whether LSN of the page is changed, then only go
for chasing all the undo chains, sure that won't work for unlogged
tables, but still it is a good optimization. Another thing is that
this approach is quite simple, so we inclined to go with this
approach.

We clear the multi_locker bit lazily like when we are already
traversing all the undo chains to verify if there is any new locker on
the tuple after taking the buffer lock.

The visibility routines don't need any special handling as we are
already storing the strongest locker information on a tuple which can
help us get the transaction information of updater which is what is
required to check visibility of tuple.

Patch by me with help from Kuntal Ghosh and Dilip Kumar, reviewed and
tested by Kuntal Ghosh