Skip unnecessary reads of newly allocated undo log pages.
Whenever we're inserting new undo data that happens to fall at the start of a
page, we know that there can be no pre-existing data on the page.  Therefore
we can ask bufmgr.c to zero it out instead of reading it from the storage
manager.

To facilitate this, create a new BufferReadMode RBM_ZERO, just like
RBM_ZERO_AND_LOCK except without the content lock.  undorecord.c expects to
acquire the content lock a bit later.  Since each backend has sole write
access to write to the undo log, it's not necessary for bufmgr.c to acquire
the lock for us.

Thomas Munro, RM43486, reviewed by XXX