Use undo-based rollback to clean up files on abort.
Use the undo log instead of "pendingDeletes" to track new files that must
be deleted if the transaction rolls back.  This way, if you kill -9 the
backend after CREATE TABLE but before COMMIT, the orphaned files will be
cleaned up after crash recovery.

This requires logging and flushing a new XLOG_SMGR_PRECREATE record
before creating relfiles, so that we know the undo records will be
recreated if we crash and recover.

Refactor undoaction.c so that it dispatches to undo handlers via
RmgrTable.  Define smgr_undo() for the orphaned file cleanup routine.
Move the code that was previously in execute_undo_actions_page() into
zheap_undo().  The prototype for undo handlers probably needs some
work.

TODO:  There is one regression test that gets an assertion failure:
rollback to savepoint where a cursor is in mid-scan, causes "buffer
is pinned in InvalidateBuffer", because we are trying to run
DropRelFileNodesAllBuffers() on a pinned buffer, when unlinking
a newly created relation.

Author: Thomas Munro