Add atomic array operations.
As a micro-optimization, allow arrays of atomic integers to be
initialized and set to zero all at once by memset(), on platforms
where that has the same effect as a loop over the regular
operations.  Provide equivalent DSA pointer operations,
and use them for Parallel Hash initialization and reset.

Author: Thomas Munro
Reviewed-by:
Discussion: