Use undo-based rollback to clean up files on abort.
Use the undo log instead of "pendingDeletes" to track new files that must
be deleted if the transaction rolls back.  This way, if you kill -9 the
backend after CREATE TABLE but before COMMIT, the orphaned files will be
cleaned up after crash recovery.

This requires logging and flushing a new XLOG_SMGR_PRECREATE record
before creating relfiles, so that we know the undo records will be
recreated if we crash and recover.

Also add a test module that can dump the contents of undo logs as NOTICE
messages, for debugging purposes.

Author: Thomas Munro
Discussion: