Allow WAL record data on first modification after a checkpoint.
Provide a way to attach data to WAL record conditionally, so that it is
included only if this turns out to the be first modification to a given
block after a checkpoint.

This will be used to record undo log meta-data, to avoid a data
synchronization problem with online checkpoints.

Author: Thomas Munro
Reviewed-by:
Discussion: