fixup! Provide interfaces to store and fetch undo records.
Patch undoinsert.c to work with the new undo log storage API.  It
has several improvements:

1.  It keeps track of transaction changes and whether the you are
at the beginning of a transaction, making life simpler for the record
insertion code.  UndoLogAllocate() tells you whether you are first
and also the address of the previous transaction in this log, so you
can decide whether to create a new transaction header and update the
previous one.  This arrangement avoids the need for a retry loop
in the code that inserts into the undo log.

2.  It no longer requires client code to deal with "meta data" and
"attach" records; they are replaced with new code that uses
registered buffers.

3.  It can now deal with a single WAL record that inserts data into
two different undo logs (because one ran out of space), simplifying
the client code (so it no longer has to allocate all undo space up
front).

The plan is for this to be squashed into the previous commit.  It
is being kept separate to make it clearer to Dilip what I needed
to change...

Author: Thomas Munro