Always perform rollback actions after cleaning up current (sub)transaction.
Currently, for error Rollbacks or when the Rollback requests are pushed to
the background worker, we perform the undo actions in a new transaction or
in parent transaction (it happens in parent transaction when we get an
error in sub-transactions).  We want the behavior to be the same for all cases
namely (a) an error, (b) rollback request executed by background worker or
(c) user-initiated rollback (to savepoint).

There are quite a few reasons to make the behavior consistent.  First one
is that if we process the undo actions in the original transaction, then we
might not get desired results for the scenario where we perform some
actions (say inserts) on a table, then altered it's tablespace and then again
performed few inserts.  We will be able to perform undo actions for the
later inserts, but not the ones which were performed before Alter tablespace.
The second case is that sometimes we hold some resources like buffer pins till
the end of the transaction and when we try to perform rollback actions for
operations where we have to forget some buffers (like drop relation files
as part of rollback), it creates problems.  The third advantage is that we
have fewer things to test.

This commit fixes both of the above problems.

One thing we have to be careful is that if we release locks before applying
undo actions, then the parallel session can acquire the lock before us
which can lead to deadlock and or possibly can start applying the actions by
itself.  So we release locks only after applying undo actions.

Amit Kapila, some review and testing by Dilip Kumar