Add XLogRecGetFullTransactionId() function.
In order to be able to work with 64 bit transaction IDs without increasing
the size of the WAL, infer the epoch.  In general we can't do that safely,
but during replay we can because we know that the next XID can't advance
concurrently (it's only advanced by replaying WAL records).