Provide interfaces to store and fetch undo records.
The undo interface layer provides a mechanism for packing/unpacking
the undo records into the undo pages.  We are assuming that the
certain operations are happening in the transaction and those
operations need to insert the undo records for handling the
visibility or for rolling back the effect of the transaction.
This layer provides interfaces for inserting and fetching the
undo records.  The patch will also provide a mechanism for
compressing the undo records.  Basically, if an undo record on
the page have the same fields value as that in the first complete
record on the page then we compress the undo record by avoid
including those field in the record provided the record is from
the same transaction as the first complete record on the page.
And, during fetch we get those informations from the first
complete record of the page.

Author: Dilip Kumar with contributions from Robert Haas, Amit Kapila,
        Thomas Munro, Vignesh C and Rafia Sabih
Reviewed-by: Earlier version of this patch is reviewed by Amit Kapila
Tested-by: Neha Sharma
Discussion: https://www.postgresql.org/message-id/CAFiTN-uVxxopn0UZ64%3DF-sydbETBbGjWapnBikNo1%3DXv78UeFw%40mail.gmail.com