Add support for discarding buffers.
To support proposed work that allows data to be truncated concurrently
with buffer access without heavyweight locking, introduce the concept
of a discarded buffer.

1.  The SMGR implementation gets a chance to report that a block has
been discarded.  This causes ReadBuffer() to return InvalidBuffer.

2.  Buffers can be discarded from the buffer pool with a new function
DiscardBuffer().  The immediately invalidates the buffer, if possible,
or marks it with a new BM_DISCARDED flag if it is already pinned, so
that it won't be written back.

Author: Thomas Munro