Handle ReadFile() end-of-file correctly on Windows.
When ReadFile() encounters the end of the file during a synchronous
operation using an offset (an operation similar to POSIX pread()),
it signals end-of-file by reporting an error instead of returning 0.
We didn't handle this correctly due to some confusion in the
documentation and the fact that it didn't break any tests, but
it generated unhelpful LOG messages about an unknown error code.
Repair.

Back-patch to 12, where we started using pread()/ReadFile() with
an offset.

Reported-by: ZhenHua Cai, Amit Kapila
Diagnosed-by: Juan Jose Santamaria Flecha
Tested-by: Amit Kapila
Discussion: https://postgr.es/m/CAA4eK1LK3%2BWRtpz68TiRdpHwxxWm%3D%2Bt1BMf-G68hhQsAQ41PZg%40mail.gmail.com