Provide infrastructure for counting background workers.
Existing infrastructure tracks the number of parallel workers
that have started and terminated.  Provide a general mechanism
to allow other kinds of workers to be counted in a similar way.
Make RegisterDynamicBackgroundWorkers() no-throw so that other
modules can reliably track workers launched, and have the
postmaster update a counter of workers terminated.