Redesign the undo bufdata encoding for closed URSs.
xactundo.c needs begin, end and the fxid whenever a URS is closed
in recovery or CLoseDanglingUndoRecordSets().  Therefore, record
a copy of the URS type-specific header with URS_XLOG_CLOSE
messages.  Generally improve the URS bufdata format.

The coding of CloseDanglingUndoRecordSets() is a bit lazy about
reacquiring buffers but it didn't seem worth the effort of avoiding
that for rare cleanup cases, and the resulting simplicity seems
to have some value.

Other small improvements:
* remove a fixed-sized buffer that arbitrarily limited header size
* improve some variable naming in passing