Introduce xid8 variants of the txid_XXX() fmgr functions.
The txid_XXX() family of functions exposes 64 bit transaction IDs to
users as int8.  Now that we have an SQL type for FullTransactionId,
define a set of functions xid8_XXX() corresponding to the txid_XXX()
functions.

It's a bit sneaky to use the same C functions for both, but since the
binary representation is identical except for the signedness of the
type, and since older functions were already using the wrong signedness,
and since we'll presumably drop the txid_XXX() variants after a period
of time, it seems reasonable to switch to FullTransactionId internally
and share the code for both.

Author: Thomas Munro
Reviewed-by: Fujii Masao, Takao Fujii, Yoshikazu Imai, Mark Dilger
Discussion: https://postgr.es/m/20190725000636.666m5mad25wfbrri%40alap3.anarazel.de