Add infrastruction for hardware transactional memory.
If --enable-htm is provided to configure, add support for using the
Intel TSX RTM facilities.  Provide a very simple abstraction that tries
once but requires the caller to provide a fallback.  In later patches,
it should be possible to map this to the builtins required for other
ISAs including ARM TME and POWER8.

Perform a runtime test at startup to check if the chip we're running on
has those instructions (eg Skylake and later server-class chips; many
mobile/laptop chips don't have them, unfortunately).

This facility will be used by later patches to perform hardware memory
transactions, to replace LWLocks and spinlocks optimistically, with a
fallback.  If compiled without --enable-htm, a constant value is provide
so that the HTM path can be removed by constant folding leaving
only the fallback path.