Use hardware transactional memory for SSI.
Replace several of the workhorse hot functions in predicate.c
with versions that first optimistically try to run in a hardware
memory transaction, but fall back to the LWLock-based versions
if the memory transaction fails.

If the optimism pays off, then we skip a whole bunch of LWLock
churn.  If it doesn't (that is, if you're frequently hitting the
same predicate lock table buckets from concurrent backends),
then it's probably slower due to frequent need to fall back
to the LWLock path, after already wasting energy on a failed
HTM transaction.

WORK IN PROGRESS -- highly experimental

Author: Thomas Munro