Rename xid to xid4.
Now that we have SQL type xid8, rename SQL type xid to xid4 for
consistency.  This makes the width of the type clearer, and avoids
creating the impression that xid is the 'real' transaction ID type,
rather than being a narrower version that lacks upper bits.

XXX Experiment only, not for commit.  To actually do this, we'd
probably want to figure out how to provide aliases like SQL xid and
C XIDOID for backwards compatibility.  It's not done in this commit,
to demonstrate that nothing in the tree depends on the old name.

XXX Documentation changes needed.