Align PGAlignedBlock to expected page size.
In order to be allowed to use O_DIRECT, we need to align buffers to the
page or sector size.  This allow on-stack buffers to be correctly
aligned.

Ideally we'd use an I/O aligned buffer for buffile.c's internal page
buffer, but for now we don't want to increase memory usage by padding
without more analysis.  Change that to use a plain old array of bytes
without special alignment.  This avoids sanitizer check failures due to
the new wider alignment attribute, but doesn't break anything because
this buffer was only ever accessed as raw bytes.

Author: Andres Freund <andres@anarazel.de>
Author: Thomas Munro <thomas.munro@gmail.com>