aio: WIP: Add POSIX AIO support.
Extend Andres Freund's PostgreSQL AIO subsystem for Linux[1] to allow
kernel-managed asynchronous I/O on the following operating systems,
when configured with --with-posix-aio:

 * FreeBSD: make check-world passing
 * NetBSD:  ?
 * macOS:   make check passing with weird delays; possibly due to known problem with unsafe code in signal handler
 * HPUX:    ?
 * AIX:     ?; manual says AIO doesn't work in mmap'd memory

To make this work on a Mac, you need to increase kern.aiomax and
kern.aiomaxproc.  That's because this patch currently panics on
submission when the kernel can't take any more requests; it needs to
learn to wait for the queue to drain.

Note: Solaris, illumos and GNU/Linux probably can't use this in the
current PostgreSQL architecture, because those operating systems have a
userland implementation of POSIX AIO based on thread pools.

Note: The infrastructure added by this commit should hopefully be
extensible to Windows IOCP.  I *guess* that'll reuse the same shared
queue implementation, but just need its own pgaio_submit_pending()
function that calls the relevant OVERLAPPED APIs, and a callback
function that replaces the SIGIO signal handler to post stuff into the
shared queue.  Design not proven yet, maybe later...

[1] https://github.com/anarazel/postgres/tree/aio