Make predictions about IO readiness to avoid syscalls.
The backend traditionally read data from the frontend until it sees
EAGAIN.  Whether or not recvfrom() is likely to succeed depends on
scheduling and workload.  Try to predict whether we should just go
directly to polling, to skip recvfrom() -> EAGAIN.

WIP:  I don't know if this makes any sense.  Probably need to check if
we're getting more than 50% EGAIN over some sample period before we
decide it's worth it!