Optionally, don't wait for end-of-recovery checkpoint.
Add a GUC to specify whether you want to wait for the end-of-recovery
checkpoint to complete before continuing.  Default to the traditional
behavior, "on".

XXX This probably needs a better GUC name.  There may also be problems with
timelines; see the nearby fast promotion code.

XXX This is experimental code only!  There may be fundamental problems
with the concept.  Or not.