Optimize setitimer() usage.
Don't call setitimer() so often.  Instead, let a preexisting alarm expire
and then request a new one as requried.  It's better to take a few extra
spurious alarm interrupts at a rate that's a function of the timeout
setting than to have a system call rate that's a function of statement
execution frequency.

XXX:  Need to review this carefully for races, perhaps involving
reentrancy.