Optimize fsync() calls in crash recovery.
Previously, we'd open every file under pgdata and call fsync(), before
entering crash recovery.  That's because it's important to avoid relying
on dirty data in the kernel's cache when deciding that a change has
already been applied.

Instead of doing work in proportion to the number of relations (which
could be extremely slow with a cold VFS cache and high latency storage),
switch to doing work in proportion to the number of relations that we
opted not to dirty (in PostgreSQL's own buffer pool) based on page LSNs.

Typically, this means doing no work at all, because most users run with
full_page_writes=on, meaning that we'll restore the earliest version of
the page and then replay every WAL record on top of it (that is, we'll
never skip a record, and always dirty, write back and eventually sync
the page).

With full_page_writes=off, we get to skip dirtying pages that have
higher LSNs ("BLK_DONE"), but now in that case we'll still queue up a
request to issue an fsync() call in the next checkpoint whenever that
happens.

Discussion: https://postgr.es/m/CAEET0ZHGnbXmi8yF3ywsDZvb3m9CbdsGZgfTXscQ6agcbzcZAw%40mail.gmail.com