Remove latch.c workaround for Linux < 2.6.27.
Ancient Linux had no XXX_CLOEXEC flags for various things, so you had to
make a separate fcntl() call.  Kernels of that vintage are long gone.
It seems like a good time to drop this extra code, because otherwise
we'd have to add similar workaround code for new patches to avoid
contradicting ourselves.