Make relfile tombstone files conditional on WAL level.
Traditionally we have left behind an empty file to be unlinked after the
next checkpoint, when dropping a relation.  That would prevent
GetNewRelFileNode() from recycling the relfile, avoiding a schedule
that could corrupt data in crash recovery, with wal_level=minimal.

Since the default wal_level changed to replica in release 10, and since
this mechanism introduces costs elsewhere, let's only do it if we have
to.  In particular, this avoids the need for DROP TABLESPACE for force
an expensive checkpoint just to clear out tombstone files.

XXX Just a sketch

XXX What would break if we did this?