Optimize empty condition variables.
Require callers of ConditionVariableBroadcast() and
ConditionVariableSignal() to make it safe for us to check if the wait
queue is empty with a simple load.  The store that will trigger any
waiter's exit condition must precede the load that will determine that
the queue is empty.

Most users of CVs already meet that requirement implicitly through the
use of LWLock, spinlock and atomic exchange operations, but one place
needs some new explicit memory barriers.