Add pg_futex_t for Linux futex.
Our traditional latches are a good general tool that supports many kinds
of waiting and multiplexing.  In some cases, we have only a limited need
for multiplexing, and we can use slightly more efficient signalling
mechanism based on modern kernel interfaces.  This commit adds
pg_futex_t and associate functions as an alternative:

  pg_futex_wait() -- wait for a shm value to change
  pg_futex_wake() -- announce that a shm value has changed

We don't want to use the futex API directly, because later patches will
also provide the same interface for other OSes that have similar but not
identical APIs.

There's also a mechanism to make sure that we can't miss a procsignal or
a postmaster death signal when we go to sleep.  This machinery will be
used by later commits to construct fast paths for latches (when not
multiplexing with sockets) and condition variables (and therefore
indirectly also for barriers, heavyweight locks, pin signals etc).

XXX WIP -- first attempt at a sketch!

XXX The signal interaction is based on incrementing the word, which may
have problems; if it just set it from 0 to 1 it'd probably be enough for
latches, but not condition variables!  There's probably a better way to
do this...