Use SIGIO to detect postmaster death.
Previously we used prctl/procctl to requests a signal on death of our
parent on Linux and FreeBSD.  Instead, use O_ASYNC to request SIGIO.
The effect is about the same, but it should work on many other Unix
systems too.  SIGIO is not in POSIX though, so we don't require it.