WIP Try to optimize buffer and lock tag lookups.
Use an inlinable, unrollable hash function for looking up buffer tags
and locks.  Take advantage of the fact that we already computed the hash
up front in these areas.

XXX Not touching LOCKLOCKTAG because they are not in shared memory and
David has other ideas for that.

XXX No results proven, yet...