Provide a new common pthread emulation for Windows.
Provide a single emulation of the subset of POSIX threads we need to
use.  This can be used to replace multiple copies found in various parts
of the tree, in follow up commits.

Increase the Windows API target to Vista, so that we can use thread
specifics with destructors (like POSIX).