Make XLogPageReadCB callbacks more flexible.
Previously, XLogPageRead() would internally wait for more data to be
streamed.  Allow it to return a special value, XLREAD_WAIT, that will be
returned by XLogReadRecord(), to indicate that the caller should sleep
or otherwise arrange for more data to be available before calling again.

For most callers, this is a mechanical change to the arguments of
XLogReadRecord(), because their installed page read function is not
capable of returning XLREAD_WAIT, and success and failure can be
detected by testing for null pointers.

For xlog.c, this minor refactoring prepares for later changes by moving
the call to WaitForWALToBecomeAvailable() outside the call stack of
XLogReadRecord().