Add fdatasync() wrapper for Windows.
Windows 10 gained support for flushing a file with fdatasync()
semantics.  It only works on NTFS local file systems, but that's
possibly OK because wal_sync_method defaults to open_datasync on that
platform.

XXX If we used pg_fdatasync() for more things in the future, we'd have
to think about how to handle the case that it doesn't work on some file
systems.

XXX This patch blithely assumes that we're on Windows 10+ without any
version checking.

XXX For experimentation only.