Make Windows FD_CLOSE reporting sticky.
Winsock reports EOF on a socket with FD_CLOSE, but only once.  Since we
have code paths that require WL_SOCKET_READABLE to be level-triggered,
we'd better remember if we'd seen FD_CLOSE and keep reporting it,
otherwise we might hang forever.  This problem was previously masked by
the fact that our sockets were not closed gracefully.

Note that this commit only fixes the problem for the lifetime of one
WaitEventSet; later commits will address the problem of losing events in
the gaps between temporary WaitEventSets.

Discussion: