Allow no-throw allocators in simplehash.h.
Previously, elog-style error reporting was assumed.  Allow SH_NO_OOM to
be defined, to indicate that you'd like creation and insertion to return
NULL to indicate failure instead, for use in no-throw programming
contexts.  A custom allocation function should return NULL on failure,
and the built-in MemoryContext support will use MCXT_NO_OOM when this
option is used.

Also allow SH_RAW_FREE to be defined, and SH_RAW_ALLOCATOR_NOZERO to
state that the SH_RAW_ALLOCATOR doesn't zero memory.  These
configuration points allow standard malloc and free to be used, which
was mostly interesting to allow testing of the SH_NO_OOM with different
allocator options.

XXX Proof-of-concept