Introduce a new abstraction for sockets.
Introduce a new type Socket that wraps sockets used in WaitEventSet.
On Unix, it simply wraps a plain old socket, but on Windows it carries
extra state required to implement reliable level-triggered events.

Specifically:

 * Each socket has its own Windows kernel event, where previously each
   WaitEventSet would create events corresponding to sockets.  The new
   arrangement allows for sockets to be used in more than one
   WaitEventSet, including temporary ones created by
   WaitLatchOrSocket().

 * Each socket remembers events already signaled by Winsock so that we
   can make them 'sticky' until the appropriate 'reset' operation is
   performed.  For example, FD_CLOSE must be reported as WL_READABLE
   forever once it has been signaled, and FD_READ must be reported as
   WL_READABLE repeatedly until the next call to pg_socket_recv().

Since some of the sockets the backend needs to wait for are constructed
inside libpq, and reconstructed at certain times without telling libpq's
client, libpq now allows its socket as a Socket.  The new interface
PQbackendsocket() is not documented and not intended as a stable API for
public use.